pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "image-multiplication"
version = "0.1.0"
description = "Image Multiplication and Download"
authors = [" Barsan Gabriel <gabrieliliebarsan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^8.0"
tk = "*"

